Split off user directory functions to dedicated module
This patch moves setup_user_data() along with mkpath() to a dedicated module.
In addition mkpath is made static as it has no callers outside of the module.
